+++ resources/views/admin/equipment-types/index.blade.php
@@ -1,6 +1,6 @@
 @extends('layouts.app')
 @section('content')
-    <x-admin.index-table title="Tipi di attrezzature"
+    <x-admin.index-table title="Tipi di attrezzature" :paginator="$equipmentTypes"
         tableClass="table table-striped table-hover my-0 align-middle text-center">
         <x-slot:headingActions>
             <x-admin.create-button :href="route('admin.equipment-types.create')" label="tipo di attrezzatura" />
@@ -17,8 +17,10 @@
             @foreach ($equipmentTypes as $equipmentType)
                 <tr>
                     <td>{{ $equipmentType->name ?? 'N/A' }}</td>
-                    <td>{{ $equipmentType->first_inspection_months ? $equipmentType->first_inspection_months . ' mesi' : 'N/A' }}</td>
-                    <td>{{ $equipmentType->regular_inspection_months ? $equipmentType->regular_inspection_months . ' mesi' : 'N/A' }}</td>
+                    <td>{{ $equipmentType->first_inspection_months ? $equipmentType->first_inspection_months . ' mesi' : 'N/A' }}
+                    </td>
+                    <td>{{ $equipmentType->regular_inspection_months ? $equipmentType->regular_inspection_months . ' mesi' : 'N/A' }}
+                    </td>
                     <x-admin.row-actions :showUrl="route('admin.equipment-types.show', $equipmentType->id)" :editUrl="route('admin.equipment-types.edit', $equipmentType->id)" :deleteTarget="'#confirmDeleteModal-' . $equipmentType->id" :label="'tipo di attrezzatura ' . $equipmentType->id" />
                 </tr>
                 <x-admin.delete-modal type="equipmentType" :object="$equipmentType" />